requirements.txt
pip==21.0.1
PyYAML==5.4.1
selenium==3.141.0
setuptools==56.0.0
urllib3==1.26.4
